% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\docType{data}
\name{.tcell_genes}
\alias{.tcell_genes}
\title{Canonical gene markers for T-cell subtypes (Zhang 2018)}
\format{
An object of class \code{list} of length 19.
}
\usage{
.tcell_genes
}
\description{
Canonical gene markers for T-cell subtypes (Zhang 2018)
}
\keyword{internal}
