% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{doublet_rate_dictionary}
\alias{doublet_rate_dictionary}
\title{Lookup expected doublet rate based on cells loaded}
\usage{
doublet_rate_dictionary(n_cells)
}
\arguments{
\item{n_cells}{Integer. Number of cells loaded onto the 10x Chromium.}
}
\value{
Numeric. Expected multiplet (doublet) rate.
}
\description{
Estimates the 10x Genomics multiplet rate (as a fraction) by finding the closest
“# of Cells Loaded” bracket in the standard guideline table.
The ratios can refer to: https://github.com/chris-mcginnis-ucsf/DoubletFinder/issues/76
We currently did not find any corresponding ratio for other single cell sequancing platforms.
}
