% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_meta_colnames}
\alias{get_meta_colnames}
\title{get colnames of srt obj meta data name}
\usage{
get_meta_colnames(srt, pattern = "RNA_snn_res")
}
\arguments{
\item{srt}{Seurat object}

\item{pattern}{Pattern to match in column names, default is "RNA_snn_res"}
}
\value{
A vector of column names that match the pattern
}
\description{
get colnames of srt obj meta data name
}
\examples{
tnk \%>\% get_meta_colnames()
}
