% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{gimme_marker}
\alias{gimme_marker}
\title{Give me marker!}
\usage{
gimme_marker(markers, cluster_of_interest = 1)
}
\arguments{
\item{markers}{A data frame or tibble containing marker genes with columns
\code{cluster}, \code{pct.1}, and \code{pct.2}.}

\item{cluster_of_interest}{The cluster number for which to find the most
distinguishing marker. Default is 1.}
}
\value{
The most distinguishing marker for the specified cluster.
}
\description{
This function gives the most distinguishing marker for a cluster
}
