% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{highlightCluster}
\alias{highlightCluster}
\title{Highlight one cluster on UMAP}
\usage{
highlightCluster(
  obj,
  cluster_id,
  reduction = "umap.harmony",
  group.by = "harmony_clusters",
  highlight.col = "red",
  other.col = "grey80",
  pt.size = 1,
  raster = TRUE,
  dpi = 72
)
}
\arguments{
\item{obj}{A Seurat object}

\item{cluster_id}{The cluster to highlight (matching levels of \code{group.by})}

\item{reduction}{Reduction slot (default "umap.harmony")}

\item{group.by}{Metadata column with cluster IDs (default "harmony_clusters")}

\item{highlight.col}{Color for the highlighted cluster (default "red")}

\item{other.col}{Color for all other clusters (default "grey80")}

\item{pt.size}{Point size (default 1)}

\item{raster}{Logical, rasterize points for speed (default TRUE)}

\item{dpi}{Rasterization DPI (default 72)}
}
\value{
ggplot2 object
}
\description{
Highlight one cluster on UMAP
}
\examples{
# Highlight cluster 34 on the harmony UMAP
highlightCluster(all, cluster_id = "34")
}
