% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\docType{data}
\name{leo.marker}
\alias{leo.marker}
\title{Marker hub}
\format{
List with two elements: \code{$data}, \code{$source}.
}
\usage{
leo.marker
}
\description{
A minimal nested list for quick access to all kinds of markers collected in \code{leo.sc}.
}
\examples{
leo.marker$t_nk_marker$data # t_nk_marker
}
\keyword{datasets}
