% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{leo.ROIE}
\alias{leo.ROIE}
\title{Compute Ro/e and draw heatmap}
\usage{
leo.ROIE(
  srt,
  filter_col = NULL,
  filter_criteria = NULL,
  anno_col = "cell_anno",
  group_col = "Stage1",
  group_col_order = NULL,
  plot = F,
  plot_path = NULL,
  col_fun = NULL,
  width = NULL,
  height = NULL,
  fontsize = 6,
  heatmap_anno = c("num", "+++", "none"),
  sym_break = c(-Inf, 0.1, 1, 2, 3, Inf),
  ...
)
}
\arguments{
\item{srt}{seurat object}

\item{filter_col}{column used to subset (optional)}

\item{filter_criteria}{values kept in \code{filter_col}}

\item{anno_col}{row group column; default \code{"cell_anno"}}

\item{group_col}{column group column; default \code{"Stage1"}}

\item{group_col_order}{column order in heatmap; auto if \code{NULL} (based on the level of \code{group_col} if it is a factor)}

\item{plot}{save heatmap if \code{TRUE}}

\item{plot_path}{PDF path; default \code{"./ROIE.pdf"}}

\item{col_fun}{colour scale: Defaut orange style or \code{redblue}}

\item{width, height}{device size; auto when \code{NULL}}

\item{fontsize}{text size inside heatmap}

\item{heatmap_anno}{\code{"num"}, \code{"+++"}, or \code{"none"}}

\item{sym_break}{numeric breakpoints; \strong{only two presets supported} --> \code{c(-Inf,.1,1,2,3,Inf)} or \code{c(-Inf,0,.2,.8,1,Inf)}}

\item{...}{passed to \code{ComplexHeatmap::Heatmap()}}
}
\value{
Ro/e matrix
}
\description{
Build a Ro/e (observed / expected) matrix from single-cell metadata and,
optionally, save a heatmap.
}
\examples{
\dontrun{
roe <- leo.ROIE(all, filter_col = 'lineage', filter_criteria = c('DC'), fontsize = 4,
                anno_col = 'cell_anno', group_col = 'Stage1', col_fun = "redblue",
                plot = T, plot_path = "./figure/sc/celltype_analysis/dc.roe.pdf",
                heatmap_anno = "+++", border = NA, rect_gp = gpar(col = NA))
head(roe)
}
}
