% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{leo_milo_vis}
\alias{leo_milo_vis}
\title{Visualize MiloR DA results}
\usage{
leo_milo_vis(
  milo_obj,
  da_results,
  plot = c(1, 2),
  layout = "UMAP.HARMONY",
  alpha = 0.05,
  log2fc_colours = c(low = "#070091", mid = "white", high = "#910000"),
  log2fc_limits = c(-5, 5),
  cell_type = "cell_anno",
  mix_threshold = 0.7,
  bee_order = NULL,
  bee_add_box = T
)
}
\arguments{
\item{milo_obj}{A \code{Milo} object.}

\item{da_results}{Output from \code{miloR::testNhoods()}.}

\item{plot}{Integer vector. Which plots to return: \code{1} for neighbourhood graph, \code{2} for beeswarm plot.}

\item{layout}{Embedding stored in the Milo object (default \code{"UMAP.HARMONY"}).}

\item{alpha}{FDR cutoff (default \code{0.05}).}

\item{log2fc_colours}{Named vector with three colours for the log2FC gradient (\code{c(low, mid, high)}).}

\item{log2fc_limits}{Two-element numeric vector giving colour-scale limits.}

\item{cell_type}{Character. Column in the metadata used to annotate neighbourhoods (default \code{"cell_anno"}).}

\item{mix_threshold}{Minimum fraction of a single annotation to avoid calling the neighbourhood “Mixed”.}

\item{bee_order}{Optional character vector specifying the display order of groups in the beeswarm plot.}

\item{bee_add_box}{Logical. If \code{TRUE}, adds a boxplot to the beeswarm plot (default \code{TRUE}).}
}
\value{
\code{list(nhood_graph, da_beeswarm)} – two \strong{ggplot} objects.
}
\description{
Visualize MiloR DA results
}
