% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{locate_most_different_g_in_2_group}
\alias{locate_most_different_g_in_2_group}
\title{Locate most distinguishing markers between two clusters}
\usage{
locate_most_different_g_in_2_group(
  sc_obj,
  ident.1,
  ident.2,
  assay = "RNA",
  test.use = "wilcox",
  pval.adj = 0.05,
  logfc = 0.5,
  min.pct1 = 0.5,
  max.pct2 = 0.2,
  return_top_n = 1
)
}
\arguments{
\item{sc_obj}{Seurat object}

\item{assay}{Assay name, default "RNA"}

\item{test.use}{DE test, one of "wilcox","roc","MAST"; default "wilcox"}

\item{pval.adj}{Adjusted p‐value cutoff, default 0.05}

\item{logfc}{|log2FC| cutoff, default 0.5}

\item{min.pct1}{Minimum detection pct in marker group, default 0.5}

\item{max.pct2}{Maximum detection pct in other group, default 0.2、}

\item{return_top_n}{Number of top markers to return for each group, default 1.}

\item{ident1}{First cluster id (string)}

\item{ident2}{Second cluster id (string)}
}
\value{
named vector c(marker1,marker2)
}
\description{
Locate most distinguishing markers between two clusters
}
\examples{
\dontrun{
tops <- locate_most_different_g_in_2_group(pbmc, "seurat_clusters", "4","5")
}
}
