% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_drop}
\alias{metadata_drop}
\title{Drop specified metadata columns from a Seurat object}
\usage{
metadata_drop(seu, ..., cols = NULL, select = NULL, ignore_missing = TRUE)
}
\arguments{
\item{seu}{Seurat object.}

\item{...}{Bare metadata names or strings (e.g., seurat_clusters, "seurat_clusters").}

\item{cols}{Optional character vector of metadata column names to drop.}

\item{select}{A tidyselect expression (e.g., starts_with("predicted")).}

\item{ignore_missing}{Logical; if TRUE, ignore missing columns and notify.}
}
\value{
Seurat object with columns removed from meta.data.
}
\description{
Drop specified metadata columns from a Seurat object
}
