% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata_get_colnames}
\alias{metadata_get_colnames}
\title{Get specified metadata columns from a Seurat object}
\usage{
metadata_get_colnames(srt, pattern = "RNA_snn_res")
}
\arguments{
\item{srt}{Seurat object.}

\item{pattern}{Pattern to match in column names, default is "RNA_snn_res".}
}
\value{
A character vector of column names that match the pattern.
}
\description{
Get specified metadata columns from a Seurat object
}
\examples{
library(Seurat)
# Create a dummy Seurat object with clustering columns
srt <- SeuratObject::pbmc_small
srt$RNA_snn_res.0.8 <- Idents(srt)
metadata_get_colnames(srt)
}
