% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{metadata_write_10x}
\alias{metadata_write_10x}
\title{Export Seurat meta.data aligned to barcodes}
\usage{
metadata_write_10x(
  seu,
  out_dir,
  file_name = "metadata.tsv.gz",
  cols = NULL,
  compress = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{seu}{Seurat object.}

\item{out_dir}{Output directory.}

\item{file_name}{Output file name.}

\item{cols}{Optional character vector of metadata columns to save.}

\item{compress}{Logical; if TRUE, write gzipped file.}

\item{overwrite}{Logical; if TRUE, overwrite existing file.}
}
\value{
Invisibly returns the output file path.
}
\description{
Export Seurat meta.data aligned to barcodes
}
