% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.plot.R
\name{plot_highlight_cluster}
\alias{plot_highlight_cluster}
\title{Plot Highlight a cluster on a dimensional reduction}
\usage{
plot_highlight_cluster(
  obj,
  cluster_id,
  reduction = NULL,
  group.by = NULL,
  highlight.col = "red",
  other.col = "grey80",
  pt.size = 1,
  dpi = 300
)
}
\arguments{
\item{obj}{A Seurat object}

\item{cluster_id}{Cluster identifier to highlight}

\item{reduction}{Dimensional reduction name (defaults to active reduction)}

\item{group.by}{Metadata column or NULL to use identities}

\item{highlight.col}{Color for highlighted cluster}

\item{other.col}{Color for other cells}

\item{pt.size}{Point size}

\item{dpi}{DPI for rasterization if >1e5 cells}
}
\value{
ggplot2 object
}
\description{
Plot Highlight a cluster on a dimensional reduction
}
