% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{plot_qc}
\alias{plot_qc}
\title{Plot QC metrics}
\usage{
plot_qc(seurat_obj, out_path, prefix = "before_")
}
\arguments{
\item{seurat_obj}{A Seurat object with QC metrics.}

\item{out_path}{Character. Directory to save plots.}

\item{prefix}{Character. Filename prefix. Default: "before_".}
}
\value{
Invisible NULL.
}
\description{
Generates violin and scatter plots for QC features and saves them as PDFs.
}
