% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{remove_unwant_hvg}
\alias{remove_unwant_hvg}
\title{Remove unwanted HVGs}
\usage{
remove_unwant_hvg(
  srt,
  pattern_list = list(mitochondrial_genes = "^MT-", ribosomal_genes = "^(RPL|RPS)",
    tcr_genes = "^TR[ABDG]", hemoglobin_genes = "^HB[AB]")
)
}
\arguments{
\item{srt}{A Seurat object with HVGs already defined
(e.g. after \code{FindVariableFeatures()}).}

\item{pattern_list}{Named list of regex patterns.  \strong{Default}
\itemize{
\item \code{mitochondrial_genes = "^MT-"}
\item \code{ribosomal_genes     = "^(RPL|RPS)"}
\item \code{tcr_genes           = "^TR[ABDG]"}
\item \code{hemoglobin_genes    = "^HB[AB]"}
}}
}
\value{
Seurat object with filtered \code{VariableFeatures()}.
A concise summary is printed.
}
\description{
Filter the current \code{VariableFeatures()} of a Seurat object and drop
genes matching user-defined regular-expression patterns. Default setting (mitochondria / ribosome / TCR / hemoglobin).
}
\examples{
\dontrun{
srt <- NormalizeData(srt) |>
       FindVariableFeatures(nfeatures = 2000) |>
       remove_unwant_hvg()
}
}
