% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype.R
\name{ROIE}
\alias{ROIE}
\title{Calculate ROIE}
\usage{
ROIE(crosstab)
}
\arguments{
\item{crosstab}{the contingency table of given distribution}
}
\value{
A matrix of Ro/e values
}
\description{
Calculate the Ro/e value from the given crosstab
}
\note{
This function is from https://github.com/yuyang3/pan-B/blob/main/Figure4.R (line 545)
}
