% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{sort_string_numeric_clusters}
\alias{sort_string_numeric_clusters}
\title{Sort string-based cluster labels numerically and refactor}
\usage{
sort_string_numeric_clusters(seurat_obj, cluster_col)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{cluster_col}{A character string. The name of the metadata column containing cluster labels as strings.}
}
\value{
The Seurat object with the specified cluster column re-factored so that its levels are sorted numerically.
}
\description{
Sort string-based cluster labels numerically and refactor
}
\examples{
# seurat_obj <- sort_string_numeric_clusters(seurat_obj, "cluster_labels")
}
