% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{subset_srt}
\alias{subset_srt}
\title{Subset srt obj easily}
\usage{
subset_srt(srt, subset_col, subset_value)
}
\arguments{
\item{srt}{Seurat object}

\item{subset_col}{Character. Column name in \code{meta.data} to subset on.}

\item{subset_value}{Character vector. Values to retain in \code{subset_col}.}
}
\value{
A subsetted Seurat object
}
\description{
Subset srt obj easily
}
