% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_meta_colnames}
\alias{get_meta_colnames}
\title{get colnames of srt obj meta data name}
\usage{
get_meta_colnames(srt, pattern = "RNA_snn_res")
}
\arguments{
\item{srt}{Seurat object}

\item{pattern}{Pattern to match in column names, default is "RNA_snn_res"}
}
\value{
A vector of column names that match the pattern
}
\description{
get colnames of srt obj meta data name
}
\examples{
library(Seurat)
# Create a dummy Seurat object with clustering columns
srt <- SeuratObject::pbmc_small
srt$RNA_snn_res.0.8 <- Idents(srt)
get_meta_colnames(srt)
}
