% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{seurat_basic_qc}
\alias{seurat_basic_qc}
\title{Seurat basic QC}
\usage{
seurat_basic_qc(seurat_obj, out_path = NULL, save_plot = T, verbose = TRUE)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{out_path}{Character. Directory to save QC plots. Required.}

\item{verbose}{Logical. Whether to print \code{leo.log} messages. Default: TRUE.}

\item{plot}{Logical. Whether to generate QC plots. Default: TRUE.}
}
\value{
A filtered Seurat object.
}
\description{
Calculates mitochondrial and hemoglobin percentages, filters cells by QC thresholds,
and optionally plots QC metrics before and after filtering.
}
