% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{seurat_basic_qc}
\alias{seurat_basic_qc}
\title{Seurat basic QC}
\usage{
seurat_basic_qc(
  seurat_obj,
  nFeature_RNA_low = 200,
  nFeature_RNA_high = 7500,
  nCount_RNA_high = 10000,
  percent_mt_high = 10,
  percent_HB_high = 1,
  out_path = NULL,
  save_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{nFeature_RNA_low}{Lower boundary for number of features per cell. Default: 200.}

\item{nFeature_RNA_high}{Upper boundary for number of features per cell. Default: 7500.}

\item{nCount_RNA_high}{Upper boundary for number of counts per cell. Default: 10000.}

\item{percent_mt_high}{Upper boundary for mitochondrial percentage. Default: 10.}

\item{percent_HB_high}{Upper boundary for hemoglobin percentage. Default: 1.}

\item{out_path}{Character. Directory to save QC plots. Required if save_plot is \code{T}.}

\item{save_plot}{Logical. Whether to save QC plots. Default: FALSE.}

\item{verbose}{Logical. Whether to print \code{leo.log} messages. Default: TRUE.}
}
\value{
A filtered Seurat object.
}
\description{
Calculates mitochondrial and hemoglobin percentages, filters cells by QC thresholds,
and optionally plots QC metrics before and after filtering.
}
