% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{seurat_standard_normalize_and_scale}
\alias{seurat_standard_normalize_and_scale}
\title{Standard seurat normalize_and_scale}
\usage{
seurat_standard_normalize_and_scale(
  seurat_obj,
  normalize_method = c("classic", "sctransform", "sctransform&regress"),
  conserve.memory = F,
  cluster_resolution = 1,
  plot_dir = NULL,
  n_hv_gene = 10,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{normalize_method}{Character. Normalization method:
\itemize{
\item "classic" (LogNormalize + FindVariableFeatures + ScaleData)
\item "sctransform" (SCTransform).
\item "sctransform&regress" (SCTransform + regress out unwanted sources \link{currently only percent.mt}).
}}

\item{conserve.memory}{Logical. Whether to conserve memory in SCTransform. Default: FALSE.}

\item{cluster_resolution}{Numeric. Resolution for FindClusters. Default: 1.0.}

\item{plot_dir}{Character. Directory to save variable feature plot. Default: NULL (i.e., not plot).}

\item{n_hv_gene}{Integer. Number of top variable features to label. Default: 10.}

\item{verbose}{Logical. Whether to print \code{leo.log} messages. Default: TRUE.}
}
\value{
A Seurat object after normalization, scaling, clustering, and UMAP embedding.
}
\description{
Performs data normalization, finds variable features, scales data, runs PCA,
constructs graph-based neighbors and clusters, and generates UMAP embeddings.
}
